Move `Parser` into a dedicated file
This commit only moves code + adds a `pub` to all fields / methods
In the future we should add tests just for parsing
